Handle station names with multibyte Unicode chars
It was causing string.format to not calculate the width of the station
name correctly, causing misalignment in station lists.  Not a big deal,
but annoying.

The solution is basically one character: use Unicode literal format
strings.  This solution seems to be hidden like a needle in an
Internet-sized haystack: https://bugs.python.org/issue7300#msg178617